requirements.txt
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
bitfinex-v2==1.6.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
httpretty==0.9.7
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyHamcrest==1.9.0
pyOpenSSL==19.1.0
requests==2.22.0
service-identity==18.1.0
six==1.13.0
Twisted==19.10.0
txaio==18.8.1
urllib3==1.25.7
zope.interface==4.7.1
